backend/requirements.txt
Flask
websockets
twitchAPI
openai
requests
black
isort
pylint~=3.0
flake8~=7.0
flake8-absolute-import~=1.0
flake8-bandit~=4.1
flake8-bugbear~=24.1
flake8-class-attributes-order~=0.1
flake8-comprehensions~=3.14
flake8-implicit-str-concat~=0.4
flake8-use-pathlib~=0.3
google-cloud-texttospeech~=2.16
